src/main/java/ru/job4j/array/JavaNameValidator.java
package ru.job4j.array;

public class JavaNameValidator {
    public static boolean isNameValid(String name) {
        if (name.isEmpty()) {
            return false;
        }
        for (int i = 0; i < name.length(); i++) {
            int code = name.codePointAt(i);
            if (i == 0 && (Character.isDigit(code) || Character.isUpperCase(code))) {
                return false;
            }
            if (!isSpecialSymbol(code) && !isUpperLatinLetter(code)
                && !isLowerLatinLetter(code) && !Character.isDigit(code)) {
                return false;
            }
        }
        return true;
    }

    public static boolean isSpecialSymbol(int code) {
        return code == 36 || code == 95;
    }

    public static boolean isUpperLatinLetter(int code) {
        return code >= 69 && code <= 90;
    }

    public static boolean isLowerLatinLetter(int code) {
        return code >= 97 && code <= 122;
    }
}
